Rename index.ts to stores.ts
